+++ example/lib/pages/image_editor_demo.dart
@@ -1,15 +1,13 @@
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:example/common/common_widget.dart';
+import 'package:example/common/crop_editor_helper.dart';
 import 'package:example/common/utils.dart';
 import 'package:example/main.dart';
 import 'package:extended_image/extended_image.dart';
 import 'package:flutter/material.dart';
 import 'package:ff_annotation_route/ff_annotation_route.dart';
-import 'package:image/image.dart';
-import 'package:image/image.dart' as image;
-import 'dart:ui' as ui;
-import 'dart:math';
 import 'package:image_picker_saver/image_picker_saver.dart';
 import 'package:oktoast/oktoast.dart';
 import 'package:image_picker/image_picker.dart' as picker;
@@ -38,6 +36,7 @@ class _ImageEditorDemoState extends State<ImageEditorDemo> {
     ..add(AspectRatioItem(text: "16*9", value: CropAspectRatios.ratio16_9))
     ..add(AspectRatioItem(text: "9*16", value: CropAspectRatios.ratio9_16));
   AspectRatioItem _aspectRatio;
+  bool _cropping = false;
   @override
   void initState() {
     _aspectRatio = _aspectRatios.first;
@@ -47,52 +46,51 @@ class _ImageEditorDemoState extends State<ImageEditorDemo> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Column(children: <Widget>[
-        AppBar(
-          title: Text("image editor demo"),
-          actions: <Widget>[
-            IconButton(
-              icon: Icon(Icons.photo_library),
-              onPressed: _getImage,
-            ),
-            IconButton(
-              icon: Icon(Icons.done),
-              onPressed: _save,
-            ),
-          ],
-        ),
-        Expanded(
-            child: _fileImage != null
-                ? ExtendedImage.file(
-                    _fileImage,
-                    fit: BoxFit.contain,
-                    mode: ExtendedImageMode.editor,
-                    enableLoadState: true,
-                    extendedImageEditorKey: editorKey,
-                    initEditorConfigHandler: (state) {
-                      return EditorConfig(
-                          maxScale: 8.0,
-                          cropRectPadding: EdgeInsets.all(20.0),
-                          hitTestSize: 20.0,
-                          initCropRectType: InitCropRectType.imageRect,
-                          cropAspectRatio: _aspectRatio.value);
-                    },
-                  )
-                : ExtendedImage.network(
-                    imageTestUrl,
-                    fit: BoxFit.contain,
-                    mode: ExtendedImageMode.editor,
-                    extendedImageEditorKey: editorKey,
-                    initEditorConfigHandler: (state) {
-                      return EditorConfig(
-                          maxScale: 8.0,
-                          cropRectPadding: EdgeInsets.all(20.0),
-                          hitTestSize: 20.0,
-                          initCropRectType: InitCropRectType.imageRect,
-                          cropAspectRatio: _aspectRatio.value);
-                    },
-                  )),
-      ]),
+      appBar: AppBar(
+        title: Text("image editor demo"),
+        actions: <Widget>[
+          IconButton(
+            icon: Icon(Icons.photo_library),
+            onPressed: _getImage,
+          ),
+          IconButton(
+            icon: Icon(Icons.done),
+            onPressed: _save,
+          ),
+        ],
+      ),
+      body: Center(
+        child: _fileImage != null
+            ? ExtendedImage.file(
+                _fileImage,
+                fit: BoxFit.contain,
+                mode: ExtendedImageMode.editor,
+                enableLoadState: true,
+                extendedImageEditorKey: editorKey,
+                initEditorConfigHandler: (state) {
+                  return EditorConfig(
+                      maxScale: 8.0,
+                      cropRectPadding: EdgeInsets.all(20.0),
+                      hitTestSize: 20.0,
+                      initCropRectType: InitCropRectType.imageRect,
+                      cropAspectRatio: _aspectRatio.value);
+                },
+              )
+            : ExtendedImage.network(
+                imageTestUrl,
+                fit: BoxFit.contain,
+                mode: ExtendedImageMode.editor,
+                extendedImageEditorKey: editorKey,
+                initEditorConfigHandler: (state) {
+                  return EditorConfig(
+                      maxScale: 8.0,
+                      cropRectPadding: EdgeInsets.all(20.0),
+                      hitTestSize: 20.0,
+                      initCropRectType: InitCropRectType.imageRect,
+                      cropAspectRatio: _aspectRatio.value);
+                },
+              ),
+      ),
       bottomNavigationBar: BottomAppBar(
         color: Colors.lightBlue,
         shape: CircularNotchedRectangle(),
@@ -114,10 +112,14 @@ class _ImageEditorDemoState extends State<ImageEditorDemo> {
                       context: context,
                       builder: (BuildContext context) {
                         return Container(
-                          height: 60.0,
-                          color: Colors.lightBlue,
-                          child: ListView.builder(
-                            scrollDirection: Axis.horizontal,
+                          //height: 60.0,
+                          //alignment: Alignment.center,
+                          //color: Colors.lightBlue.withOpacity(0.4),
+                          child: GridView.builder(
+                            gridDelegate:
+                                SliverGridDelegateWithFixedCrossAxisCount(
+                                    crossAxisCount: 3),
+                            padding: EdgeInsets.all(20.0),
                             itemBuilder: (_, index) {
                               var item = _aspectRatios[index];
                               return GestureDetector(
@@ -191,32 +193,36 @@ class _ImageEditorDemoState extends State<ImageEditorDemo> {
     );
   }
 
+  Uint8List data;
   void _save() async {
+    if (_cropping) return;
+    var msg = "";
     try {
-      var cropRect = editorKey.currentState.getCropRect();
-      ui.Image imageData = editorKey.currentState.image;
+      _cropping = true;
 
-      var data = await imageData.toByteData(format: ui.ImageByteFormat.png);
-      image.Image src = decodePng(data.buffer.asUint8List());
+      showBusyingDialog();
 
-      if (editorKey.currentState.editAction.hasEditAction) {
-        var editAction = editorKey.currentState.editAction;
-        src = copyFlip(src, flipX: editAction.flipX, flipY: editAction.flipY);
-        if (editAction.hasRotateAngle) {
-          double angle = (editAction.rotateAngle ~/ (pi / 2)) * 90.0;
-          src = copyRotate(src, angle);
-        }
-      }
+      ///delay due to cropImageDataWithDartLibrary is time consuming on main thread
+      ///it will block showBusyingDialog
+      ///if you don't want to block ui, use compute/isolate,but it costs more time.
+      await Future.delayed(Duration(milliseconds: 200));
 
-      var cropData = copyCrop(src, cropRect.left.toInt(), cropRect.top.toInt(),
-          cropRect.width.toInt(), cropRect.height.toInt());
+      ///if you don't want to block ui, use compute/isolate,but it costs more time.
+      var fileData =
+          await cropImageDataWithDartLibrary(state: editorKey.currentState);
+      // var fileData =
+      //     await cropImageDataWithNativeLibrary(state: editorKey.currentState);
 
-      var fileFath =
-          await ImagePickerSaver.saveFile(fileData: encodePng(cropData));
-      showToast("save image : $fileFath");
+      var fileFath = await ImagePickerSaver.saveFile(fileData: fileData);
+      msg = "save image : $fileFath";
     } catch (e) {
-      showToast("save faild: $e");
+      msg = "save faild: $e";
+      print(msg);
     }
+
+    Navigator.of(context).pop();
+    showToast(msg);
+    _cropping = false;
   }
 
   File _fileImage;
@@ -225,25 +231,39 @@ class _ImageEditorDemoState extends State<ImageEditorDemo> {
         await picker.ImagePicker.pickImage(source: picker.ImageSource.gallery);
 
     setState(() {
+      editorKey.currentState.reset();
       _fileImage = image;
     });
   }
-}
 
-image.Image copyFlip(image.Image src,
-    {bool flipX = false, bool flipY = false}) {
-  if (!flipX && !flipY) return src;
-
-  image.Image dst = image.Image(src.width, src.height,
-      channels: src.channels, exif: src.exif, iccp: src.iccProfile);
-
-  for (int yi = 0; yi < src.height; ++yi,) {
-    for (int xi = 0; xi < src.width; ++xi,) {
-      var sx = flipY ? src.width - 1 - xi : xi;
-      var sy = flipX ? src.height - 1 - yi : yi;
-      dst.setPixel(xi, yi, src.getPixel(sx, sy));
-    }
+  Future showBusyingDialog() async {
+    var primaryColor = Theme.of(context).primaryColor;
+    return showDialog(
+        context: context,
+        barrierDismissible: false,
+        child: Material(
+          color: Colors.transparent,
+          child: Container(
+            height: double.infinity,
+            width: double.infinity,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                // CircularProgressIndicator(
+                //   strokeWidth: 2.0,
+                //   valueColor:
+                //       AlwaysStoppedAnimation(primaryColor),
+                // ),
+                // SizedBox(
+                //   width: 10.0,
+                // ),
+                Text(
+                  "cropping...",
+                  style: TextStyle(color: primaryColor),
+                )
+              ],
+            ),
+          ),
+        ));
   }
-
-  return dst;
 }