+++ lib/src/editor/extended_image_editor.dart
@@ -1,6 +1,8 @@
 import 'dart:math';
+import 'dart:typed_data';
 import 'package:extended_image/src/extended_image_utils.dart';
 import 'package:extended_image/src/image/extended_raw_image.dart';
+import 'package:extended_image_library/extended_image_library.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'extended_image_crop_layer.dart';
@@ -17,6 +19,8 @@ class ExtendedImageEditor extends StatefulWidget {
   ExtendedImageEditor({this.extendedImageState, Key key})
       : assert(extendedImageState.imageWidget.fit == BoxFit.contain,
             "Make sure the image is all painted to crop,the fit of image must be BoxFit.contain"),
+        assert(extendedImageState.imageWidget.image is ExtendedImageProvider,
+            "Make sure the image provider is ExtendedImageProvider, we will get raw image data from it"),
         super(key: key);
   @override
   ExtendedImageEditorState createState() => ExtendedImageEditorState();
@@ -245,16 +249,22 @@ class ExtendedImageEditorState extends State<ExtendedImageEditor> {
       return null;
     }
 
+    var cropScreen = _editActionDetails.screenCropRect;
     var imageScreenRect = _editActionDetails.screenDestinationRect;
-    var cropScreen =
-        _editActionDetails.screenCropRect.shift(-imageScreenRect.topLeft);
+    imageScreenRect = _editActionDetails.paintRect(imageScreenRect);
+    cropScreen = _editActionDetails.paintRect(cropScreen);
+
+    //move to zero
+    cropScreen = cropScreen.shift(-imageScreenRect.topLeft);
+
     imageScreenRect = imageScreenRect.shift(-imageScreenRect.topLeft);
 
     var image = widget.extendedImageState.extendedImageInfo.image;
-    var size = _editActionDetails.isHalfPi
-        ? Size(image.height.toDouble(), image.width.toDouble())
-        : Size(image.width.toDouble(), image.height.toDouble());
-    var imageRect = Offset.zero & size;
+    // var size = _editActionDetails.isHalfPi
+    //     ? Size(image.height.toDouble(), image.width.toDouble())
+    //     : Size(image.width.toDouble(), image.height.toDouble());
+    var imageRect =
+        Offset.zero & Size(image.width.toDouble(), image.height.toDouble());
 
     var ratioX = imageRect.width / imageScreenRect.width;
     var ratioY = imageRect.height / imageScreenRect.height;
@@ -268,6 +278,10 @@ class ExtendedImageEditorState extends State<ExtendedImageEditor> {
   }
 
   ui.Image get image => widget.extendedImageState.extendedImageInfo?.image;
+
+  Uint8List get rawImageData =>
+      (widget.extendedImageState?.imageWidget?.image as ExtendedImageProvider)
+          .rawImageData;
 
   EditActionDetails get editAction => _editActionDetails;