+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.6.3]
+
+* Issues:
+  Fix issue that forget canvas.restore after canvas.clipRect
+* Breaking Change:
+  ImageEditor：you should crop image data before flip or rotate image data now.
+* Improve:
+  Increase cropping speed
+
 ## [0.6.2]
 
 * Features: